refactor: update component names
