+++ src/components/AccountSelect/AccountItem.style.tsx
@@ -1,8 +1,6 @@
 import styled from 'styled-components';
 
-import { Flex } from '@/component-library/Flex';
-import { Span } from '@/component-library/Text';
-import { theme } from '@/component-library/theme';
+import { Flex, Span, theme } from '@/component-library';
 
 type StyledListItemSelectedLabelProps = {
   $isSelected: boolean;
